refactor(P2P): moved transport into its own package
