+++ p2p/transport/transport.go
@@ -1,4 +1,4 @@
-package p2p
+package transport
 
 import (
 	"fmt"
